% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatingModel.R
\name{calculate_single_Index}
\alias{calculate_single_Index}
\title{Function for integrating observation models}
\usage{
calculate_single_Index(IndexObj)
}
\arguments{
\item{IndexObj}{A Index object}
}
\description{
Function for integrating observation models
}
