% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatingModel.R
\name{calculate_single_LengthComp}
\alias{calculate_single_LengthComp}
\title{Function for integrating length composition observation models}
\usage{
calculate_single_LengthComp(dataObject)
}
\arguments{
\item{LengthCompObj}{A Length Comp observation model object}
}
\description{
Function for integrating length composition observation models
}
