% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatingModel.R
\name{calculateIndex}
\alias{calculateIndex}
\title{Observation models function for testing}
\usage{
calculateIndex(simulation_result, IndexObj)
}
\arguments{
\item{simulation_result}{Output from runProjection}

\item{IndexObj}{A Index object}
}
\description{
Observation models function for testing
}
