% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managementStrategies.R
\name{constantCatchStrategy}
\alias{constantCatchStrategy}
\title{MP for setting a TAC based on average catch of final x years of historical.}
\usage{
constantCatchStrategy(phase, dataObject)
}
\arguments{
\item{phase}{Management procedures are coded in three phases: 1 - data collection, 2 - a decision making process, 3 - conversion of that process into annual F}

\item{dataObject}{The needed inputs to the management procedure}
}
\description{
The Strategy object should be specified as follows. (1) Set Strategy@projectionYears to the number of forward projection years you wish to simulate. (2) Strategy@projectionName = "constantCatchStrategy".
(3)  Strategy@projectionParams should be a named list with one item called aveYrs, where aveYrs is the number of years to average catch across, including and proceeding the final historical year.
}
