% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatingModel.R
\name{cpueDev}
\alias{cpueDev}
\title{Initial cpue deviations}
\usage{
cpueDev(TimeAreaObj, StochasticObj = NULL)
}
\arguments{
\item{TimeAreaObj}{A TimeArea object}

\item{StochasticObj}{A Stochastic object}
}
\description{
Initial cpue deviations
}
