% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{ensure_matrix_format}
\alias{ensure_matrix_format}
\title{Internal helper function to ensure observation data is in matrix format}
\usage{
ensure_matrix_format(simulation_result)
}
\arguments{
\item{simulation_result}{Output object from runProjection()}
}
\value{
simulation_result with observation data in matrix format
}
\description{
Internal helper function to ensure observation data is in matrix format
}
\keyword{internal}
