% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG.R
\docType{package}
\name{fishSimGTG}
\alias{fishSimGTG}
\alias{fishSimGTG-package}
\title{Numerical simulations of fish population dynamics}
\description{
Conducts projections, MSE, and equilibrium conditions for a specified fish life history.
}
\author{
\strong{Maintainer}: William Harford \email{bill@natureanalytics.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
