% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{get_metric_units}
\alias{get_metric_units}
\title{Extract units from simulation result}
\usage{
get_metric_units(simulation_result)
}
\arguments{
\item{simulation_result}{Output from runProjection}
}
\value{
List of units for different metrics
}
\description{
Extract units from simulation result
}
\keyword{internal}
