% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatingModel.R
\name{LHwrapper}
\alias{LHwrapper}
\title{Life history wrapper for sub-cohorts}
\usage{
LHwrapper(LifeHistoryObj, TimeAreaObj)
}
\arguments{
\item{LifeHistoryObj}{A life history object.}

\item{TimeAreaObj}{A time-area object}
}
\description{
Creates the necessary age-based vectors describing life history of sub-cohorts
}
\examples{
ta<-new("TimeArea")
ta@gtg<-13
ta@stepsPerYear<-1
LHwrapper(LifeHistoryObj = LifeHistoryExample, TimeAreaObj=ta)
}
