% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesS4.R
\docType{class}
\name{LifeHistory-class}
\alias{LifeHistory-class}
\title{Life history object}
\arguments{
\item{title}{A title for the object, useful for displaying the contents of the object}

\item{speciesName}{Scientific name of the species}

\item{shortDescription}{A brief description of the object. This could be the common name, stock, geographic location of the stock, etc.}

\item{L_type}{The method of measuring length. e.g. TL for total length. Must be consistent for all length params e.g., Linf, L50, L95}

\item{L_units}{Units of measure for the object. cm is expected. Must be consistent for all length params e.g., Linf, L50, L95}

\item{W_units}{Units of weight for the object. Must be consistent with LW_A and LW_B params}

\item{Linf}{von Bertalanffy Loo parameter}

\item{K}{von Bertalanffy K parameter per year}

\item{t0}{von Bertalanffy t0 parameter}

\item{L50}{Length at 50\% maturity.}

\item{L95}{Length at 95\% maturity. Must be a value larger than L50}

\item{M}{Natural mortality rate per year}

\item{MK}{Ratio of M to K}

\item{LW_A}{Parameter for length-weight relationship W=aL^b}

\item{LW_B}{Parameter for length-weight relationship W=aL^b}

\item{Tmax}{Maximum observed age}

\item{Steep}{Steepness of the Beverton-Holt stock recruit relationship}

\item{R0}{Unfished number of recruits}

\item{recSD}{Inter annual recruitment variation}

\item{recRho}{Inter-annual correlation in recruitment variation}

\item{isHermaph}{Logical whether species is a protogynous hermaphrodite (TRUE). FALSE is gonochoristic species.}

\item{H50}{Length at 50\% male.}

\item{H95}{Length at 95\% male. Must be a value larger than H50}

\item{author}{Author of the life history - do not include your email address as these life histories may be posted online}

\item{authAffiliation}{A way to identify the author without relying on an email address}

\item{longDescription}{Document the rational for choices made in creating the life history. Other users will rely on this information.}

\item{appBuild}{A data frame that holds details when the life history is built using the Shiny app}

\item{ID}{A user ID. Not required.}
}
\description{
An S4 object that holds a description of a life history.
}
\details{
This S4 object is used as to input a life history.
}
