% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{plot_all_indices}
\alias{plot_all_indices}
\title{Plot all indices}
\usage{
plot_all_indices(simulation_result, ...)
}
\arguments{
\item{simulation_result}{Output object from \code{runProjection()} containing observation data}

\item{...}{Additional arguments passed to \code{plot_indices()}}
}
\value{
A ggplot object with faceted panels for each index
}
\description{
Creates time series plots of all indices (both survey and CPUE) from fishSimGTG observation models.
This is a wrapper function that includes all index types.
}
\seealso{
\code{\link{plot_indices}}
}
