% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{plot_catch_observations_single}
\alias{plot_catch_observations_single}
\title{Plot catch observations (single fleet mode)}
\usage{
plot_catch_observations_single(simulation_result, ...)
}
\arguments{
\item{simulation_result}{Output object from \code{runProjection()} containing catch observation data}

\item{...}{Additional arguments passed to \code{plot_catch_observations()}}
}
\value{
A ggplot object showing aggregated catch time series
}
\description{
Creates time series plots of catch observations using single fleet aggregation.
Forces single fleet display even in multifleet simulations.
}
\seealso{
\code{\link{plot_catch_observations}}
}
