% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{plot_catchB}
\alias{plot_catchB}
\title{Plot catch biomass}
\usage{
plot_catchB(result, areas = "all", ...)
}
\arguments{
\item{result}{Simulation result from runProjection}

\item{areas}{Areas to plot ("all" or numeric vector)}

\item{...}{Additional arguments passed to plot_population_metric}
}
\description{
Plot catch biomass
}
