% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{plot_cpue_indices}
\alias{plot_cpue_indices}
\title{Plot CPUE indices}
\usage{
plot_cpue_indices(simulation_result, ...)
}
\arguments{
\item{simulation_result}{Output object from \code{runProjection()} containing observation data}

\item{...}{Additional arguments passed to \code{plot_indices()}}
}
\value{
A ggplot object with faceted panels for each CPUE index
}
\description{
Creates time series plots of fishery-dependent CPUE indices from fishSimGTG observation models.
This is a wrapper function that filters for CPUE indices (IDX_CPUE pattern).
}
\seealso{
\code{\link{plot_indices}}
}
