% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{plot_discB}
\alias{plot_discB}
\title{Plot dicard catch biomass}
\usage{
plot_discB(result, ...)
}
\arguments{
\item{result}{Simulation result from runProjection}

\item{...}{Additional arguments passed to plot_population_metric}

\item{areas}{Areas to plot ("all" or numeric vector)}
}
\description{
Plot dicard catch biomass
}
