% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{plot_fishery_length_comp_modified}
\alias{plot_fishery_length_comp_modified}
\title{Wrapper for fishery length compositions}
\usage{
plot_fishery_length_comp_modified(
  result,
  area_filter = "all",
  fleet_filter = "all",
  ...
)

plot_fishery_length_comp_modified(
  result,
  area_filter = "all",
  fleet_filter = "all",
  ...
)
}
\description{
Wrapper for fishery length compositions

Wrapper for fishery length compositions - modified (proportions version)
}
