% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{plot_length_comp_modified}
\alias{plot_length_comp_modified}
\title{Modified Length Compositions (using proportions)}
\usage{
plot_length_comp_modified(
  simulation_result,
  fleet_id = 1,
  years = "all",
  show_median = TRUE,
  show_quantiles = TRUE,
  show_individual = FALSE,
  color_palette = NULL,
  title = NULL,
  save_plot = FALSE,
  filename = NULL,
  width = 14,
  height = 10
)
}
\arguments{
\item{simulation_result}{Output from runProjection}

\item{fleet_id}{Fleet number to plot}

\item{years}{Years to plot or "all"}

\item{show_median}{Show median line}

\item{show_quantiles}{Show quantile ribbon}

\item{show_individual}{Show individual iterations}

\item{color_palette}{Custom colors}

\item{title}{Custom title}

\item{save_plot}{Save to file}

\item{filename}{Custom filename}

\item{width}{Plot width}

\item{height}{Plot height}
}
\description{
Enhanced version with larger fonts and proportions instead of raw counts
}
