% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{plot_recN}
\alias{plot_recN}
\title{Plot recruitment}
\usage{
plot_recN(result, ...)
}
\arguments{
\item{result}{Simulation result from runProjection}

\item{...}{Additional arguments passed to plot_population_metric}
}
\description{
Plot recruitment
}
