% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{plot_recN_modified}
\alias{plot_recN_modified}
\title{Modified Recruitment}
\usage{
plot_recN_modified(
  simulation_result,
  show_median = TRUE,
  show_quantiles = TRUE,
  show_individual = FALSE,
  color_palette = NULL,
  title = NULL,
  save_plot = FALSE,
  filename = NULL,
  width = 14,
  height = 10
)
}
\description{
Enhanced version with larger fonts and standardization
}
