% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{plot_SPR}
\alias{plot_SPR}
\title{Plot spawning potential ratio}
\usage{
plot_SPR(result, ...)
}
\arguments{
\item{result}{Simulation result from runProjection}

\item{...}{Additional arguments passed to plot_population_metric}
}
\description{
Plot spawning potential ratio
}
