% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{plot_survey_indices}
\alias{plot_survey_indices}
\title{Plot survey indices}
\usage{
plot_survey_indices(simulation_result, ...)
}
\arguments{
\item{simulation_result}{Output object from \code{runProjection()} containing observation data}

\item{...}{Additional arguments passed to \code{plot_indices()}}
}
\value{
A ggplot object with faceted panels for each survey index
}
\description{
Creates time series plots of fishery-independent survey indices from fishSimGTG observation models.
This is a wrapper function that filters for survey indices (IDX_Survey pattern).
}
\seealso{
\code{\link{plot_indices}}
}
