% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{plot_TAC}
\alias{plot_TAC}
\title{Plot TAC (Total Allowable Catch) by area}
\usage{
plot_TAC(
  simulation_result,
  areas = "all",
  show_fleets = FALSE,
  show_median = TRUE,
  show_quantiles = TRUE,
  show_individual = FALSE,
  title = NULL,
  save_plot = FALSE,
  filename = NULL
)
}
\arguments{
\item{simulation_result}{Output from runProjection}

\item{areas}{Character "all" or numeric vector}

\item{show_fleets}{Logical. Show fleet breakdown in multifleet}

\item{...}{Additional plotting arguments}
}
\description{
Plot TAC (Total Allowable Catch) by area
}
