% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{plot_TAC_by_area}
\alias{plot_TAC_by_area}
\title{Plot TAC by area (wrapper)}
\usage{
plot_TAC_by_area(result, areas = "all", ...)
}
\arguments{
\item{result}{Simulation result}

\item{areas}{Areas to plot}

\item{...}{Additional arguments}
}
\description{
Plot TAC by area (wrapper)
}
