% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{plot_TAC_by_fleet}
\alias{plot_TAC_by_fleet}
\title{Plot TAC by fleet (wrapper)}
\usage{
plot_TAC_by_fleet(result, ...)
}
\arguments{
\item{result}{Simulation result}

\item{...}{Additional arguments}
}
\description{
Plot TAC by fleet (wrapper)
}
