% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{plot_TAC_total_modified}
\alias{plot_TAC_total_modified}
\title{Modified Total TAC by Fleet}
\usage{
plot_TAC_total_modified(
  simulation_result,
  show_fleets = TRUE,
  show_median = TRUE,
  show_quantiles = TRUE,
  show_individual = FALSE,
  color_palette = NULL,
  title = NULL,
  save_plot = FALSE,
  filename = NULL,
  width = 14,
  height = 10
)
}
\arguments{
\item{simulation_result}{Output from runProjection}

\item{show_fleets}{Show fleet breakdown}

\item{show_median}{Show median line}

\item{show_quantiles}{Show quantile ribbon}

\item{show_individual}{Show individual iterations}

\item{color_palette}{Custom colors}

\item{title}{Custom title}

\item{save_plot}{Save to file}

\item{filename}{Custom filename}

\item{width}{Plot width}

\item{height}{Plot height}
}
\description{
Enhanced version with larger fonts and standardization by catch
Uses HCR$decisionAnnual like the original
}
