% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatingModel.R
\name{plotIndex_simple}
\alias{plotIndex_simple}
\title{Plot function for observation models for testing}
\usage{
plotIndex_simple(
  index_result,
  save_plot = FALSE,
  filename = "index_simple.jpeg"
)
}
\arguments{
\item{index_result}{Output from index Object}

\item{save_plot}{saving plot}

\item{filename}{file name}
}
\description{
Plot function for observation models for testing
}
