% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEWrappers.R
\name{runProjection}
\alias{runProjection}
\title{Run the projection or MSE model}
\usage{
runProjection(
  LifeHistoryObj,
  TimeAreaObj,
  HistFisheryObj,
  ProFisheryObj_list = NULL,
  StrategyObj = NULL,
  StochasticObj = NULL,
  MultifleetObj = NULL,
  IndexObj = NULL,
  CatchObsObj = NULL,
  LengthCompObj = NULL,
  wd,
  fileName,
  seed = 1,
  doPlot = FALSE,
  doDiagnostic = F,
  customToCluster = NULL,
  titleStrategy = "No name",
  waitName = NULL,
  hostName = NULL
)
}
\arguments{
\item{LifeHistoryObj}{A LifeHistory object. Required}

\item{TimeAreaObj}{A TimeArea object. Required}

\item{HistFisheryObj}{A Fishery object that characterizes the historical dynamics. Required as it is used in initial equilibrium and historical time dynamics (if applicable)}

\item{ProFisheryObj_list}{A Fishery object used in forward projection. Optional, only used when StrategyObj is supplied}

\item{StrategyObj}{A Strategy object. Optional}

\item{StochasticObj}{A Stochastic object. Optional}

\item{wd}{A working directly to save output. Required}

\item{fileName}{A file name for output. Required}

\item{seed}{A value used in base::set.seed function for producing consistent set of stochastic elements. Optional}

\item{doPlot}{Logical whether to produce diagnostic plots upon completing simulations. Default is FALSE (no plots)}

\item{customToCluster}{A character vector containing name or names of custom management strategies to export to the cluster (otherwise parallel processing will fail).}

\item{titleStrategy}{A title for management strategy being evaluated.}

\item{waitName}{When used within a shiny app, this function can update a host from the waiter package. See example.}

\item{hostName}{When used within a shiny app, this function can update a host from the waiter package. See example.}
}
\description{
Function for running projections or MSE
}
