% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managementStrategies.R
\name{solveFfromTAC}
\alias{solveFfromTAC}
\title{Calculate F from TAC - utility function.}
\usage{
solveFfromTAC(dataObject)
}
\arguments{
\item{dataObject}{The needed inputs to the management procedure}
}
\description{
Calculate F from TAC - utility function.
}
