% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managementStrategies.R
\name{solveTAC_to_F_fishSimGTG}
\alias{solveTAC_to_F_fishSimGTG}
\title{Function to convert Total Allowable Catch (TAC) targets into F rates for multiple fleets using Newton-Raphson iteration, adapted from OpenMSE and https://api.admb-project.org/baranov_8cpp_source.html code for fishSimGTG}
\usage{
solveTAC_to_F_fishSimGTG(
  j,
  k,
  TAC_targets,
  N,
  lh,
  selGroup,
  M_rate,
  effort_F_by_fleet,
  is_multifleet,
  areas,
  nfleets,
  TAC_type = "keep",
  control = NULL
)
}
\arguments{
\item{j}{Current simulation year}

\item{k}{Current iteration}

\item{TAC_targets}{Vector of TAC targets \link{fleet1_TAC, fleet2_TAC, ...}. NA = effort-managed}

\item{N}{List of abundance arrays by GTG: N[\link{gtg}][age, year, area]}

\item{lh}{Life history object from LHwrapper}

\item{selGroup}{Selectivity object: selGroup[\link{area}][\link{fleet}] (multifleet) or selGroup[\link{area}] (single fleet)}

\item{M_rate}{Natural mortality rate (scalar)}

\item{effort_F_by_fleet}{Vector of F values for effort-managed fleets (length = nfleets)}

\item{is_multifleet}{Logical indicating multifleet mode}

\item{areas}{Number of areas}

\item{nfleets}{Number of fleets (1 for single fleet)}

\item{TAC_type}{Character: "keep" (retained catch) or "removals" (total removals)}

\item{control}{List with algorithm settings (maxiterF, tolF)}
}
\value{
Vector of F values \link{fleet1_F, fleet2_F, ...} or single value for single fleet
}
\description{
Function to convert Total Allowable Catch (TAC) targets into F rates for multiple fleets using Newton-Raphson iteration, adapted from OpenMSE and https://api.admb-project.org/baranov_8cpp_source.html code for fishSimGTG
}
