% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesS4.R
\docType{class}
\name{TimeArea-class}
\alias{TimeArea-class}
\title{Time-area object}
\arguments{
\item{title}{A title for the object, useful for displaying the contents of the object}

\item{gtg}{Number of growth-type groups}

\item{stepsPerYear}{Number of time steps per year (e.g., 1 for annual time step, 12 for monthly time step)}
}
\description{
An S4 object that holds descriptions of time step, gtg, and area params
}
\details{
This S4 object is used as input to a variety of functions, including various population dynamics simulation functions.
Options for vulnerability and retention retention functions along with guidance on parameter specification is found in Sel documentation
}
