% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatingModel.R
\name{calculate_single_CatchObs}
\alias{calculate_single_CatchObs}
\title{Function for integrating observation models}
\usage{
calculate_single_CatchObs(dataObject)
}
\arguments{
\item{CatchObsObj}{A Catch observation model object}
}
\description{
Function for integrating observation models
}
