% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG.R
\docType{package}
\name{fishSimGTG}
\alias{fishSimGTG-package}
\alias{fishSimGTG}
\title{Numerical simulations of fish population dynamics}
\description{
Conducts projections, MSE, and equilibrium conditions for a specified fish life history.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natureanalytics-ca/fishSimGTG}
  \item Report bugs at \url{https://github.com/natureanalytics-ca/fishSimGTG/issues}
}

}
\author{
\strong{Maintainer}: William Harford \email{bill@natureanalytics.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
