% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatingModel.R
\name{histEffortDev}
\alias{histEffortDev}
\title{Historical effort deviations}
\usage{
histEffortDev(TimeAreaObj, StochasticObj)
}
\arguments{
\item{TimeAreaObj}{A TimeArea object}

\item{StochasticObj}{A Stochastic object}
}
\description{
Historical effort deviations
}
