% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatingModel.R
\name{LHwrapper}
\alias{LHwrapper}
\title{Life history wrapper for sub-cohorts}
\usage{
LHwrapper(LifeHistoryObj)
}
\arguments{
\item{LifeHistoryObj}{A life history object.}
}
\description{
Creates the necessary age-based vectors describing life history of sub-cohorts
}
\examples{
LHwrapper(LifeHistoryExample)
}
