% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesS4.R
\docType{class}
\name{Multifleet-class}
\alias{Multifleet-class}
\title{Multifleet object}
\arguments{
\item{title}{A title for the object}

\item{nfleets}{Number of fleets}

\item{fleet_proportions}{Numeric vector of fleet proportions (must sum to 1)}

\item{allocation_type}{Either "effort" or "catch" for equilibrium allocation}

\item{fleet_selectivity_hist_list}{List of Fishery objects (length = nfleets)}

\item{fleet_selectivity_proj_list}{List of Fishery objects (length = nfleets)}
}
\description{
An S4 object that holds multifleet configuration for historical and projection periods
}
