% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{plot_fishery_length_comp_modified}
\alias{plot_fishery_length_comp_modified}
\title{Wrapper for fishery length compositions}
\usage{
plot_fishery_length_comp_modified(result, areas = "all", fleets = "all", ...)
}
\description{
Wrapper for fishery length compositions
}
