% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{plot_Ftotal}
\alias{plot_Ftotal}
\title{Plot total fishing mortality}
\usage{
plot_Ftotal(result, areas = "all", ...)
}
\arguments{
\item{result}{Simulation result from runProjection}

\item{areas}{Areas to plot ("all" or numeric vector)}

\item{...}{Additional arguments passed to plot_population_metric}
}
\description{
Plot total fishing mortality
}
