% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{plot_indices_modified}
\alias{plot_indices_modified}
\title{Modified Indices Plots (Survey and CPUE)}
\usage{
plot_indices_modified(
  simulation_result,
  index_pattern = "IDX_",
  show_median = TRUE,
  show_quantiles = TRUE,
  show_individual = FALSE,
  point_size = 2,
  line_alpha = 0.5,
  color_palette = NULL,
  title = NULL,
  save_plot = FALSE,
  filename = NULL,
  width = 14,
  height = 10
)
}
\arguments{
\item{simulation_result}{Output from runProjection}

\item{index_pattern}{Pattern to match (e.g., "IDX_Survey_1", "IDX_CPUE_2")}

\item{show_median}{Show median line}

\item{show_quantiles}{Show quantile ribbon}

\item{show_individual}{Show individual iterations}

\item{point_size}{Size of points}

\item{line_alpha}{Transparency}

\item{color_palette}{Custom colors}

\item{title}{Custom title}

\item{save_plot}{Save to file}

\item{filename}{Custom filename}

\item{width}{Plot width}

\item{height}{Plot height}
}
\description{
Enhanced version with larger fonts, keeps original flexible pattern matching
}
