% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{plot_survey_length_comp_modified}
\alias{plot_survey_length_comp_modified}
\title{Wrapper for survey length compositions}
\usage{
plot_survey_length_comp_modified(result, areas = "all", ...)
}
\description{
Wrapper for survey length compositions
}
