% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishSimGTG-plot-functions.R
\name{plot_VB}
\alias{plot_VB}
\title{Plot vulnerable biomass}
\usage{
plot_VB(result, areas = "all", ...)
}
\arguments{
\item{result}{Simulation result from runProjection}

\item{areas}{Areas to plot ("all" or numeric vector)}

\item{...}{Additional arguments passed to plot_population_metric}
}
\description{
Plot vulnerable biomass
}
